requirements.txt
Flask==2.3.0
vosk==0.3.45
pyaudio==0.2.13
requests==2.31.0
google-generativeai>=0.8.3